Small final refactoring on rest module
