+++ opensilex-core/src/main/java/org/opensilex/core/experiment/api/ExperimentAPI.java
@@ -25,7 +25,7 @@ import org.opensilex.core.experiment.dal.ExperimentModel;
 import org.opensilex.server.response.ErrorResponse;
 import org.opensilex.server.response.ObjectUriResponse;
 import org.opensilex.server.response.SingleObjectResponse;
-import org.opensilex.rest.security.ApiProtected;
+import org.opensilex.rest.authentication.ApiProtected;
 import org.opensilex.sparql.service.SPARQLService;
 import org.opensilex.sparql.exceptions.SPARQLAlreadyExistingUriException;
