+++ opensilex-dev-tools/src/main/java/org/opensilex/dev/Install.java
@@ -44,8 +44,8 @@ import org.eclipse.rdf4j.rio.RDFParser;
 import org.eclipse.rdf4j.rio.Rio;
 import org.eclipse.rdf4j.rio.helpers.StatementCollector;
 import org.opensilex.OpenSilex;
-import org.opensilex.rest.security.AuthenticationService;
-import org.opensilex.rest.security.dal.UserDAO;
+import org.opensilex.rest.authentication.AuthenticationService;
+import org.opensilex.rest.user.dal.UserDAO;
 import org.opensilex.sparql.service.SPARQLService;
 import org.opensilex.sparql.rdf4j.RDF4JConfig;
 import org.opensilex.utils.ClassUtils;
