+++ opensilex-rest/src/main/java/org/opensilex/rest/security/dal/GroupDAO.java
@@ -9,7 +9,7 @@ import java.net.URI;
 import java.util.List;
 import org.apache.jena.arq.querybuilder.SelectBuilder;
 import org.apache.jena.sparql.expr.Expr;
-import org.opensilex.rest.security.SecurityOntology;
+import org.opensilex.rest.authentication.SecurityOntology;
 import org.opensilex.sparql.service.SPARQLQueryHelper;
 import org.opensilex.sparql.service.SPARQLService;
 import org.opensilex.sparql.utils.OrderBy;
