+++ opensilex-rest/src/main/java/org/opensilex/rest/security/dal/SecurityProfileModel.java
@@ -5,7 +5,7 @@
  */
 package org.opensilex.rest.security.dal;
 
-import org.opensilex.rest.security.SecurityOntology;
+import org.opensilex.rest.authentication.SecurityOntology;
 import java.util.List;
 import org.apache.jena.vocabulary.DCTerms;
 import org.opensilex.sparql.annotations.SPARQLProperty;
